1.6.0 Support Linux Using LibRealSense 2.10.0

+++ include/_librealsense2/rs_advanced_mode.hpp
@@ -270,7 +270,7 @@ namespace rs400
             rs2_error* e = nullptr;
             rs2_load_json(_dev.get(),
                           json_content.data(),
-                          (unsigned int)json_content.size(),
+                          json_content.size(),
                           &e);
             rs2::error::handle(e);
         }
